% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chinook.R
\name{addTab,Chinook-method}
\alias{addTab,Chinook-method}
\title{add a ChinookTab subclass to this Chinook instance}
\usage{
\S4method{addTab}{Chinook}(obj, tab)
}
\arguments{
\item{obj}{A Chinook object}

\item{tab}{A ChinookTab object}
}
\value{
nothing
}
\description{
add a ChinookTab subclass to this Chinook instance
}
