% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chinook.R
\docType{class}
\name{Chinook}
\alias{Chinook}
\alias{.Chinook}
\title{import shiny}
\usage{
Chinook(name, quiet = TRUE)
}
\arguments{
\item{name}{A character string}
}
\value{
An object of the Chinook class
}
\description{
a shiny app
}
