% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChinookTab.R
\docType{class}
\name{ChinookTab}
\alias{ChinookTab}
\alias{.ChinookTab}
\title{import shiny
import cyjShiny
import graph}
\usage{
ChinookTab(name, quiet = TRUE)
}
\arguments{
\item{namem}{A character string}
}
\value{
An object of the ChinookTab class
}
\description{
a shiny app
}
