% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chinook.R
\name{createServer,Chinook-method}
\alias{createServer,Chinook-method}
\alias{createServer}
\title{Create the shiny server}
\usage{
\S4method{createServer}{Chinook}(obj, session, input, output)
}
\arguments{
\item{obj}{An object of class TrenaViz}

\item{session}{A shiny session}

\item{input}{A shiny input object}

\item{output}{A shiny output object}
}
\description{
Create the shiny server
}
