% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chinook.R
\name{createUI,Chinook-method}
\alias{createUI,Chinook-method}
\alias{createUI}
\title{Create the user interface}
\usage{
\S4method{createUI}{Chinook}(obj)
}
\arguments{
\item{obj}{An object of class Chinook}
}
\description{
Create the user interface
}
