% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoTabOne.R
\name{DemoTabOne}
\alias{DemoTabOne}
\title{Create a DemoTabOne object}
\usage{
DemoTabOne(name, menuItemName, parentApp, quiet = TRUE)
}
\arguments{
\item{namem}{A character string}
}
\value{
An DemoTabOne object, a subclass of Chinookab
}
\description{

}
