% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoTabTwo.R
\name{DemoTabTwo}
\alias{DemoTabTwo}
\title{Create a DemoTabTwo object}
\usage{
DemoTabTwo(name, menuItemName, parentApp, quiet = TRUE)
}
\arguments{
\item{namem}{A character string}
}
\value{
An DemoTabTwo object, a subclass of Chinookab
}
\description{

}
