% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChinookTab.R
\name{getMenuItemName,ChinookTab-method}
\alias{getMenuItemName,ChinookTab-method}
\alias{getMenuItemName}
\title{get the menu item name associated with this tab}
\usage{
\S4method{getMenuItemName}{ChinookTab}(obj)
}
\arguments{
\item{obj}{An object of class TrenaViz}
}
\value{
character string
}
\description{
get the menu item name associated with this tab
}
