% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChinookTab.R
\name{getName,ChinookTab-method}
\alias{getName,ChinookTab-method}
\alias{getName}
\title{get the name of this tab}
\usage{
\S4method{getName}{ChinookTab}(obj)
}
\arguments{
\item{obj}{An object of class TrenaViz}
}
\value{
character string
}
\description{
get the name of this tab
}
