% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chinook.R
\name{getTabs,Chinook-method}
\alias{getTabs,Chinook-method}
\title{get all the tabs associated with this Chinook instance}
\usage{
\S4method{getTabs}{Chinook}(obj)
}
\arguments{
\item{obj}{A Chinook objet}
}
\value{
A list of ChinookTab objects
}
\description{
get all the tabs associated with this Chinook instance
}
