% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chinook.R
\name{Chinook}
\alias{Chinook}
\alias{printf}
\title{import shiny
import jsonlite}
\usage{
printf(...)

Chinook(name, homePage.htmlFile = NA, quiet = TRUE)
}
\arguments{
\item{name}{A character string}
}
\value{
An object of the Chinook class
}
\description{
a shiny app
}
